Update emoji table
The emoji table is the essence of why there are even READMEs. People decide if they want to use the software based on the choice of emojis and how appropriate there are. That's why we have to refine our choices, to become better, until we reach the limit of what you can do with emoji tables.

GitHub emoji table preview is broken (basically not usable now), so you'll have to diff using the raw view